requirements.txt
pyimagej>=1.4.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0
openjdk-11
